refactor(web): split submissions table components
