+++ apps/web/src/routes/admin/submissions/submissions-filter-option-list.svelte
@@ -1,0 +1,60 @@
+<script lang="ts">
+  import { ScrollArea, Spinner } from '$lib/components'
+  import SubmissionsFilterOption from './submissions-filter-option.svelte'
+  import SubmissionsFilterSearchInput from './submissions-filter-search-input.svelte'
+  import type { ValueFilterFamily } from './submissions-filter-ui'
+
+  interface Props {
+    family: ValueFilterFamily
+    searchable?: boolean
+    mobile?: boolean
+  }
+
+  let { family, searchable = false, mobile = false }: Props = $props()
+  const options = $derived(family.options())
+</script>
+
+{#snippet optionList()}
+  <div class={mobile ? 'flex flex-col gap-1 p-2' : 'p-1'}>
+    {#if family.loading?.()}
+      <div
+        class={mobile
+          ? 'text-foreground-l3 flex items-center gap-2 px-2 py-8 text-sm'
+          : 'text-foreground-l3 flex items-center gap-2 px-2 py-1.5 text-sm'}
+      >
+        <Spinner class="size-3.5" />
+        {family.loadingLabel}
+      </div>
+    {:else if options.length === 0}
+      <div
+        class={mobile
+          ? 'text-foreground-l3 px-2 py-8 text-center text-sm'
+          : 'text-foreground-l3 px-2 py-1.5 text-sm'}
+      >
+        {family.emptyLabel}
+      </div>
+    {:else}
+      {#each options as option (family.optionKey(option))}
+        <SubmissionsFilterOption {family} {option} {mobile} />
+      {/each}
+    {/if}
+  </div>
+{/snippet}
+
+{#if searchable && family.search}
+  <SubmissionsFilterSearchInput
+    value={family.search.value()}
+    placeholder={family.search.placeholder}
+    onInput={family.search.onInput}
+  />
+  <ScrollArea
+    class="min-h-0 flex-1"
+    fadeSize={28}
+    fadeColor="background-l4"
+    scrollbarYClasses="hidden"
+  >
+    {@render optionList()}
+  </ScrollArea>
+{:else}
+  {@render optionList()}
+{/if}
